profile/cpu: Generalized CPUParser
